Use the release_ue function to send a delayed RRC Connection Release
This fully fixes pcap logging.

When not delaying the sending of the RRC Connection Release, the
attach reject message is not being properly logged in the enb pcap
dump. Not delaying on the cleanup, means we might be cleaning up the
connection too early. We could have also done two defer callbacks:
sending RRC Connection Release (50ms) and after that cleaning
up (2*50ms=100ms). However, this is ugly. It's best to just delay the
release_ue function which already has a delay for the cleanup in
process_release_complete.